Corrected lookup for rho
In the standard implementation the rho used in the turbulence equations was a geometricOneField rather than the actual density. The model have now been updated to include the actual density. The change has little impact on the results as the buoyancy production term effectively separated the two phases (this bug was only present in the v1712 and OF50 versions. In extend, which was also used in the article the actual density was used)